requirements.txt
# Vercel Python Runtime Requirements
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
curl-cffi==0.7.4
dataclasses-json==0.6.7
fastapi==0.115.6
filetype==1.2.0
frozendict==2.4.6
frozenlist==1.5.0
google-ai-generativelanguage==0.6.13
google-api-core==2.25.0
google-api-python-client==2.158.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.7
googleapis-common-protos==1.67.0
greenlet==3.1.1
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-google-genai==2.1.2
langchain-text-splitters==0.3.4
langgraph==0.2.72
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.2.0
langgraph-sdk==0.1.48
langsmith==0.2.14
marshmallow==3.25.0
multidict==6.1.0
multitasking==0.0.11
mypy-extensions==1.0.0
numpy==2.2.1
orjson==3.10.14
ormsgpack==1.5.0
packaging==24.2
pandas==2.2.3
peewee==3.17.8
pinecone==5.4.0
pinecone-plugin-assistant==1.2.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.6
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic-core==2.27.2
pyparsing==3.2.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.3
tenacity==9.0.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.1.0
typing-extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==14.1
xxhash==3.5.0
yarl==1.18.3
yfinance==0.2.50
zstandard==0.23.0
